Broke Camera Butttt.....
So Camera's broken when you go towards the 2nd level but I think I just gotta fix the Red Dead Space Gizmo and make it shorter so that the camera goes down. I downloaded all the scripts and other prefabs/assets for platformers. Just a matter of dropping the things into the game.